all_requirements.txt
# Core visualization & dashboard
plotly==5.20.0
dash==2.16.1
dash-bootstrap-components
dash-core-components
dash-ag-grid
dash-image-gallery
dash-holoniq-wordcloud
dash-cytoscape
wordcloud

# Data and utils
pandas
umap-learn
scikit-learn
wget
numpy==2.0.2
networkx==3.2.1
psutil==7.0.0
packaging==25.0
regex==2024.11.6
scipy==1.13.1
PyYAML==6.0.2
requests==2.32.3
typing_extensions==4.14.0
importlib_metadata==8.7.0
zipp==3.22.0
Jinja2==3.1.6
MarkupSafe==3.0.2
click==8.1.8
filelock==3.18.0
charset-normalizer==3.4.2
idna==3.10
certifi==2025.4.26
urllib3==2.4.0
sympy==1.14.0
mpmath==1.3.0
PyWavelets==1.6.0
invisible-watermark==0.2.0

# Metrics 
bert-score==0.3.13
lpips
brisque==0.0.16
scikit-image>=0.22

# OpenAI & ML
openai
git+https://github.com/openai/CLIP.git
transformers==4.52.4
tokenizers==0.21.1
huggingface-hub==0.32.4
hf-xet==1.1.3
fsspec==2025.5.1
accelerate==1.7.0
diffusers==0.33.1
safetensors==0.5.3

# PyTorch stack
torch==2.7.1
torchvision

# CUDA (for GPU)
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
triton==3.3.1

# Image
pillow==11.2.1
opencv-python==4.11.0.86

# Jupyter
notebook
ipykernel

# dotenv
python-dotenv